eval: Include diagnostics in judge prompt (#28990)
We will now include diagnostics before and after the change in the judge
prompt.

We also added a check to prevent inadvertently including an example with
preexisting diagnostics. This might actually be desired to see whether
model can discern them from the ones it introduced, so examples in this
state can be run by setting `allow_preexisting_diagnostics = true`.

If the example configured to not run a language server, we'll explicitly
indicate this in the judge prompt.

Release Notes:

- N/A

---------

Co-authored-by: Thomas Mickley-Doyle <tmickleydoyle@gmail.com>
Co-authored-by: Bennet Bo Fenner <bennetbo@gmx.de>