Fix zed sometimes stopping by using setsid on interactive shell
For some reason `SIGTTIN` sometimes gets sent to the process group, causing it to stop when run from a terminal. This solves that issue by putting the shell in a new session + progress group.

This allows removal of a workaround of using `exit 0;` to restore handling of ctrl-c after exit.  In testing this appears to no longer be necessary.

Closes #27716