agent: Reject shouldn't delete AI-created file with user edits
Closes https://github.com/zed-industries/zed/issues/30527

The fix prevents data loss, but it results in a somewhat confusing UX.
Specifically, when the user makes changes to an AI-created file,
selecting "Reject" will leave AI changes in place. This is because
there's no easy way to distinguish between edits made by the user and
edits made by the AI (or I didn't find it).

A better solution would be to find a way to revert AI changes only. In
the mean time, this change should do.